Add MPFR, GMP and MPC to the MSP430 toolchain
Using specific versions for MPFR, GMP and MPC makes the gcc not to depend on
the system-wide versions. The build is more hermetic this way and allows the
same deb packet to be used on both Ubuntu Lucid and Debian Squeeze.

The versions used are the following:

  - mpfr: 2.4.2
  - gmp: 4.3.2
  - mpc: 0.8.1

These are the oldest versions recommended for gcc 4.5.